web/src/requirements/base.txt
django==3.2.10
django-filter==2.4.0

# for PostgreSql database
psycopg2==2.9.1

# Ddos defender: required redis
django-defender==0.9.1

# redis
redis==3.5.3

# for RestAPI
djangorestframework==3.13.1
drf-yasg==1.20.0 # Swagger

# for auth system
django-allauth==0.42.0
dj-rest-auth==2.1.12
djangorestframework-simplejwt==5.0.0

# For backend working
celery==5.2.1
pyamqp==0.1.0.7

# for react
django-cors-headers==3.10.1

# For testing
flake8==4.0.1

# Async Support
uvicorn[standard]==0.16.0

# Cashe Support
django-cacheops==6.0
django-redis==5.1.0

# ImageField
pillow==8.1.1

# Translation
django-rosetta==0.9.8

# Monitoring System
sentry-sdk==1.5.0

# MicroService Requests
django-microservice-request==0.1.4.8

# RangeFilter for admin
django-admin-rangefilter==0.8.2

# JsonValidator
jsonschema==4.2.1
